Solucion de bug al cargar pantalla de carga Modifición en la busqueda Metodos para filtros perfil, fichaje, edad.

+++ BusDep.ViewModel/JugadorBusquedaViewModel.cs
@@ -47,7 +47,7 @@ namespace BusDep.ViewModel
         public DateTime? FechaNacimiento { get; set; }
 
         [DataMember]
-        public long? Edad {
+        public int? Edad {
             get
             {
                 if(!FechaNacimiento.HasValue)
